requirements.txt
asgiref==3.3.1
autopep8==1.5.5
coverage==5.4
crispy-tailwind==0.2.0
Django==3.1.6
django-countries==7.0
django-crispy-forms==1.10.0
django-extensions==3.1.0
django-filter==2.4.0
django-meta==2.0.0
django-model-utils==4.1.1
django-tinymce==3.2.0
djangorestframework==3.12.2
flake8==3.8.4
graphviz==0.16
mccabe==0.6.1
Pillow==8.1.0
pycodestyle==2.6.0
pydot==1.4.1
pydot-ng==2.0.0
pyflakes==2.2.0
pyparsing==2.4.7
pytz==2021.1
six==1.15.0
sqlparse==0.4.1
toml==0.10.2
